requirements.txt
requests
PyYAML
markdownify
pydantic>=2.0
aiohttp
aiofiles
tqdm
colorama>=0.4.6
